refactor: invfx frame을 각각 나눔
코드를 더 효율적으로 관리할 수 invfx의 frame을 각각 나눔
패키지명을 단수로 변경
public 클래스를 모두 internal 클래스로 변경
