+++ src/main/kotlin/io/github/teamuselessplugin/punishment/event/BlockEvents.kt
@@ -1,4 +1,4 @@
-package io.github.teamuselessplugin.punishment.events
+package io.github.teamuselessplugin.punishment.event
 
 import net.kyori.adventure.text.Component
 import org.bukkit.GameMode
@@ -18,7 +18,7 @@ import org.bukkit.event.player.PlayerMoveEvent
 import java.util.*
 import kotlin.collections.HashMap
 
-class BlockEvents : Listener {
+internal class BlockEvents : Listener {
     companion object {
         val tracking: HashMap<UUID, Boolean> = HashMap()
         val trackingPlayer: HashMap<UUID, UUID> = HashMap()
