+++ src/main/kotlin/io/github/teamuselessplugin/punishment/interface/Command.kt
@@ -1,0 +1,6 @@
+package io.github.teamuselessplugin.punishment.`interface`
+
+internal interface Command {
+    val commandName: String
+    fun push()
+}
