--- src/main/kotlin/io/github/teamuselessplugin/punishment/interfaces/Command.kt
@@ -1,6 +1,0 @@
-package io.github.teamuselessplugin.punishment.interfaces
-
-interface Command {
-    val commandName: String
-    fun push()
-}
